Small code simplification for counting coefficients
replaced repeated computation of those permutation coefficients with a single function housed in utils. Baesd on a simple version of cvec_2_pvec(), which may be slower. We could plug in cvec_2_pvec() in case that would give performance gains.  Includes unit test to verify that the two methods are equivalent.